backend/v1/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
authlib==1.6.0
itsdangerous==2.1.2
cryptography==41.0.7  # For Fernet encryption (PII protection)

# HTTP Client
httpx==0.25.2

# Cache (Optional)
redis==5.0.1

# Testing (for future use)
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
python-dotenv==1.0.0